+++ gradio/image_utils.py
@@ -29,12 +29,18 @@ def format_image(
     elif type == "filepath":
         try:
             path = processing_utils.save_pil_to_cache(
-                im, cache_dir=cache_dir, name=name, format=fmt or format  # type: ignore
+                im,
+                cache_dir=cache_dir,
+                name=name,
+                format=fmt or format,  # type: ignore
             )
         # Catch error if format is not supported by PIL
         except (KeyError, ValueError):
             path = processing_utils.save_pil_to_cache(
-                im, cache_dir=cache_dir, name=name, format="png"  # type: ignore
+                im,
+                cache_dir=cache_dir,
+                name=name,
+                format="png",  # type: ignore
             )
         return path
     else:
@@ -54,7 +60,9 @@ def save_image(y: np.ndarray | _Image.Image | str | Path, cache_dir: str):
         fmt = y.format
         try:
             path = processing_utils.save_pil_to_cache(
-                y, cache_dir=cache_dir, format=fmt  # type: ignore
+                y,
+                cache_dir=cache_dir,
+                format=fmt,  # type: ignore
             )
         # Catch error if format is not supported by PIL
         except (KeyError, ValueError):
@@ -71,3 +79,26 @@ def save_image(y: np.ndarray | _Image.Image | str | Path, cache_dir: str):
         )
 
     return path
+
+
+def crop_scale(img: _Image.Image, final_width: int, final_height: int):
+    original_width, original_height = img.size
+    target_aspect_ratio = final_width / final_height
+
+    if original_width / original_height > target_aspect_ratio:
+        crop_height = original_height
+        crop_width = crop_height * target_aspect_ratio
+    else:
+        crop_width = original_width
+        crop_height = crop_width / target_aspect_ratio
+
+    left = (original_width - crop_width) / 2
+    top = (original_height - crop_height) / 2
+
+    img_cropped = img.crop(
+        (int(left), int(top), int(left + crop_width), int(top + crop_height))
+    )
+
+    img_resized = img_cropped.resize((final_width, final_height))
+
+    return img_resized